requirements.txt
cffi==1.5.2
pyalsaaudio==0.8.2
pycparser==2.14
pyspotify==2.0.5
RPi.GPIO==0.6.1
SQLAlchemy==1.0.12
psycopg2==2.6.1
